tests/requirements.txt
black
flake8
model-bakery
pytest
pytest-django
tox
tox-gh-actions
tox-pyenv
